pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meshshare"
version = "0.1.0"
description = "Mesh-aware file sharing and peer discovery web server for Linux hosts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116,<1",
  "uvicorn[standard]>=0.35,<1",
  "pydantic>=2.11,<3",
  "pyyaml>=6.0,<7",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1",
  "pytest>=8.4,<9",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
